+++ CHANGES.md
@@ -11,6 +11,7 @@ Release Notes.
 * Add witness class/method for resteasy-server plugin(v3/v4/v6)
 * Add async-profiler feature for performance analysis 
 * Support db.instance tag,db.collection tag and AggregateOperation span for mongodb plugin(3.x/4.x)
+* Improve CustomizeConfiguration by avoiding repeatedly resolve file config
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/222?closed=1)